pyproject.toml
[project]
name = "e-pharmacy-backend"
version = "0.1.0"
description = "Production grade backend for e pharmacy commerce that supports prescription validation and secure order workflows"
authors = [
    {name = "koladefaj", email = "fajimikolade23@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.13"

# PRODUCTION dependencies only
dependencies = [
    "fastapi>=0.128.0,<0.129.0",
    "uvicorn>=0.40.0,<0.41.0",
    "sqlalchemy>=2.0.45,<3.0.0",
    "alembic>=1.18.0,<2.0.0",
    "asyncpg>=0.31.0,<0.32.0",
    "psycopg2-binary>=2.9.11,<3.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "celery>=5.6.2,<6.0.0",
    "slowapi>=0.1.9,<0.2.0",
    "httpx>=0.28.1,<0.29.0",
    "pydantic[email]>=2.12.5,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "redis>=7.1.0,<8.0.0",
    "boto3>=1.42.27,<2.0.0",
    "python-magic>=0.4.27; sys_platform != 'win32'",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "asgiref>=3.11.0,<4.0.0",
    "google>=3.0.0,<4.0.0",
    "google-genai>=1.57.0,<2.0.0",
    "python-multipart (>=0.0.22,<0.0.23)",
    "stripe>=14.2.0,<15.0.0",
    "sendgrid>=6.12.5,<7.0.0",
    "reportlab>=4.4.9,<5.0.0",
    "certifi>=2026.1.4,<2027.0.0",
    "ecdsa (>=0.19.1,<0.20.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
]

[tool.poetry.group.dev.dependencies]
# DEVELOPMENT dependencies only
pytest = ">=9.0.2,<10.0.0"
pytest-asyncio = ">=1.3.0,<2.0.0"
black = ">=25.12.0,<26.0.0"
isort = ">=7.0.0,<8.0.0"
bandit = ">=1.9.3,<2.0.0"
ruff = ">=0.1.0,<0.2.0"  # Add ruff for linting
aiosqlite = ">=0.22.1,<0.23.0"
python-magic-bin = { version = ">=0.4.14,<0.5.0", markers = "sys_platform == 'win32'" }

[tool.poetry.group.test.dependencies]
# Optional: separate test dependencies if needed
pytest-cov = ">=4.1.0,<5.0.0"
pytest-mock = ">=3.11.1,<4.0.0"

[tool.poetry]
package-mode = false

[tool.isort]
profile = "black"
line_length = 88

[[tool.poetry.source]]
name = "safety"
url = "https://pkgs.safetycli.com/repository/none-19107/project/e-pharmacy-backend/pypi/simple/"
priority = "primary"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]